pyproject.toml
[tool.poetry]
name = "nonebot-plugin-mcsmanager"
version = "0.0.1"
description = "MCSManager plugin for nonebot"
authors = ["kexue <x@kexue.io>"]
readme = "README.md"
license = "MIT"

[tool.poetry.dependencies]
python = "^3.7.3"
httpx = "^0.20.0"
nonebot2 = "^2.0.0.beta.1"
nonebot-adapter-onebot = "^2.0.0-beta.1"
# nonebot-plugin-htmlrender = "^0.0.4"

[tool.poetry.dev-dependencies]
black = {version = "^22.1.0", allow-prereleases = true}
isort = "^5.10.1"
mypy = "^0.931"
flake8 = "^3"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
